Moved files to Programs folder and linked to 6 week program
